+++ packages/ai/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @lunora/ai [1.0.0-alpha.55](https://github.com/anolilab/lunora/compare/%40lunora%2Fai%401.0.0-alpha.54...%40lunora%2Fai%401.0.0-alpha.55) (2026-08-10)
+
+
+### Dependencies
+
+* **@lunora/errors:** upgraded to 1.0.0-alpha.20
+
 ## @lunora/ai [1.0.0-alpha.54](https://github.com/anolilab/lunora/compare/%40lunora%2Fai%401.0.0-alpha.53...%40lunora%2Fai%401.0.0-alpha.54) (2026-08-10)