+++ packages/scheduler/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @lunora/scheduler [1.0.0-alpha.28](https://github.com/anolilab/lunora/compare/%40lunora%2Fscheduler%401.0.0-alpha.27...%40lunora%2Fscheduler%401.0.0-alpha.28) (2026-08-10)
+
+
+### Dependencies
+
+* **@lunora/errors:** upgraded to 1.0.0-alpha.20
+
 ## @lunora/scheduler [1.0.0-alpha.27](https://github.com/anolilab/lunora/compare/%40lunora%2Fscheduler%401.0.0-alpha.26...%40lunora%2Fscheduler%401.0.0-alpha.27) (2026-08-10)