+++ packages/server/CHANGELOG.md
@@ -1,3 +1,12 @@
+## @lunora/server [1.0.0-alpha.69](https://github.com/anolilab/lunora/compare/%40lunora%2Fserver%401.0.0-alpha.68...%40lunora%2Fserver%401.0.0-alpha.69) (2026-08-10)
+
+
+### Dependencies
+
+* **@lunora/errors:** upgraded to 1.0.0-alpha.20
+* **@lunora/scheduler:** upgraded to 1.0.0-alpha.28
+* **@lunora/values:** upgraded to 1.0.0-alpha.25
+
 ## @lunora/server [1.0.0-alpha.68](https://github.com/anolilab/lunora/compare/%40lunora%2Fserver%401.0.0-alpha.67...%40lunora%2Fserver%401.0.0-alpha.68) (2026-08-09)