+++ packages/values/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @lunora/values [1.0.0-alpha.25](https://github.com/anolilab/lunora/compare/%40lunora%2Fvalues%401.0.0-alpha.24...%40lunora%2Fvalues%401.0.0-alpha.25) (2026-08-10)
+
+
+### Dependencies
+
+* **@lunora/errors:** upgraded to 1.0.0-alpha.20
+
 ## @lunora/values [1.0.0-alpha.24](https://github.com/anolilab/lunora/compare/%40lunora%2Fvalues%401.0.0-alpha.23...%40lunora%2Fvalues%401.0.0-alpha.24) (2026-08-10)