+++ packages/x402/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @lunora/x402 [1.0.0-alpha.35](https://github.com/anolilab/lunora/compare/%40lunora%2Fx402%401.0.0-alpha.34...%40lunora%2Fx402%401.0.0-alpha.35) (2026-08-10)
+
+
+### Dependencies
+
+* **@lunora/errors:** upgraded to 1.0.0-alpha.20
+
 ## @lunora/x402 [1.0.0-alpha.34](https://github.com/anolilab/lunora/compare/%40lunora%2Fx402%401.0.0-alpha.33...%40lunora%2Fx402%401.0.0-alpha.34) (2026-08-09)